Added setting screen
